Implement RoomSettings
- Implement Packet_RequestTransfer
- Implement Packet_Room
- Implement Packet_Host
- Implement Packet_Shop
- Implement Packet_Lobby
- Implement Packet_RoomList

+++ networking/tcp_server.h
@@ -38,7 +38,7 @@ private:
 	}
 
 	IPV _ipVersion;
-	int _port;
+	unsigned short _port;
 
 	thread _serverThread;
 	boost::asio::io_context _ioContext;
